requirements.txt
appdirs==1.4.4
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.13.1
fonttools==4.44.3
fsspec==2023.10.0
ftfy==6.1.1
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
install==1.3.5
Jinja2==3.1.2
kiwisolver==1.4.5
lmdb==1.4.1
loguru==0.7.2
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
Pillow==10.1.0
protobuf==4.25.1
pyarrow==14.0.1
pycocotools==2.0.7
pyparsing==3.1.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sentry-sdk==1.35.0
setproctitle==1.3.3
smmap==5.0.1
sympy==1.12
torch==1.12.0+cu116
torchaudio==0.12.0+cu116
torchvision==0.13.0+cu116
tqdm==4.66.1
triton==2.1.0
urllib3==2.1.0
wandb==0.16.0
